backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Shun Yuan <tanshunyuan99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.1"
marshmallow-sqlalchemy = "^0.21.0"
psycopg2 = "^2.8.4"
flask-cors = "^3.0.8"
flask-sqlalchemy = "^2.4.1"
flask-migrate = "^2.5.2"
loguru = "^0.4.1"
bcrypt = "^3.1.7"
python-dotenv = "^0.10.5"
sqlalchemy-utils = "^0.36.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
